backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-generativeai
SQLAlchemy
alembic
passlib[bcrypt]
python-jose[cryptography]
python-multipart
slowapi